refactor: Move routes to /api
