+++ src/api/routes/health/consts.rs
