+++ src/api/routes/v1/error.rs
