+++ src/api/routes/v1/friendships/errors.rs
