+++ src/api/routes/v1/friendships/get.rs
@@ -9,7 +9,7 @@ use crate::{
     components::{app::AppComponents, synapse::clean_synapse_user_id},
     entities::friendships::{Friendship, FriendshipRepositoryImplementation},
     middlewares::check_auth::UserId,
-    routes::v1::error::CommonError,
+    api::routes::v1::error::CommonError,
 };
 
 const ME: &str = "me";
