+++ src/api/routes/v1/friendships/types.rs
