+++ src/lib.rs
@@ -2,23 +2,23 @@ pub mod components;
 pub mod entities;
 mod metrics;
 pub mod middlewares;
-pub mod routes;
+pub mod api;
 mod utils;
 
 use actix_web::body::MessageBody;
 use actix_web::dev::{Server, ServiceFactory};
-use actix_web::middleware::{self};
+use actix_web::middleware;
 use actix_web::{web::Data, App, HttpServer};
 use middlewares::check_auth::CheckAuthToken;
-use routes::synapse::room_events::room_event_handler;
-use routes::v1::friendships::mutuals::get_mutual_friends;
+use api::routes::synapse::room_events::room_event_handler;
+use api::routes::v1::friendships::mutuals::get_mutual_friends;
 use tracing_actix_web::TracingLogger;
 
 use components::{app::AppComponents, configuration::Config, tracing::init_telemetry};
 use metrics::initialize_metrics;
 use middlewares::metrics_token::CheckMetricsToken;
-use routes::v1::friendships::get::get_user_friends;
-use routes::{
+use api::routes::v1::friendships::get::get_user_friends;
+use api::routes::{
     health::handlers::{health, live},
     synapse::handlers::{login, version},
 };
@@ -30,8 +30,6 @@ pub struct AppOptions {
 
 pub fn run_service(data: Data<AppComponents>) -> Result<Server, std::io::Error> {
     init_telemetry();
-
-    log::debug!("App Config: {:?}", data.config);
 
     let server_host = data.config.server.host.clone();
     let server_port = data.config.server.port;
